requirements.txt
# Core ML and Data
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Telegram Bot
python-telegram-bot==20.7

# LLM Integration
groq==0.4.1

# Utilities
python-dotenv==1.0.0
faker==20.1.0
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Development
pytest==7.4.3
black==23.11.0